requirements.txt
pymongo==4.10.1
psycopg2-binary==2.9.10
pandas==2.1.4
faker==33.3.0
apache-airflow-providers-mongo==4.2.2
sqlmodel==0.0.6